[Enhancement] Export participantslist to Excel #364
Two API methods:

GET /api/v1/registrations/export
GET /api/v1/registrations/export/{eventId}

+ integration tests.

IMPL DETAILS

First, new service is created called IRegistrationRetrievalService.
There's no good way currently to extend the CRUD operations which doesn't
violate one of SOLID principles, added comment about that in IRegistrationService.

So IRegistrationRetrievalService is an example of how it can be done in a flexible
way without violating the common principles. Also, this service returns Paging which
is also a new way of retrieving data. With that paging, you just cannot select more
than 100 records at a time. This should  be a good shelter against possible memory
overflow bugs and a various DDOS attacks.

To retrieve more than 100 records one should loop through all the pages. The workaround
for this is calling PageReader.ReadAllAsync which may be called  in special cases
(like in export which is the current task).

ALSO IN THIS COMMIT

.editorconfig is modified to accept private members with underscore prefix +
to allow private members to be accessed without adding 'this.'. This is the current
code style, see no reason why do we need to support boths styles in one source tree.