feat: add Image  component
This commit adds a new component called Image. The Image component is used to display images with optional caption and custom dimensions. It accepts props such as `src`, `alt`, `caption`, `className`, `width`, and `height`.

The commit also includes updates to the `index.ts` file to expose the Image component.

Lastly, the commit adds a placeholder image
 file under the `public/images` directory.