Feature/eventcreatesplit (#117)
* feat: separate event creation from event editing

Splits event editor in two new components: EventCreator and EventEditor. The EventCreator component handles the creation of new events, while the EventEditor component is responsible for editing existing events.

* docs: adds docs on AppNotifications

The commit updates the Storybook configuration file to include .mdx files in the stories and adds a new .mdx file for documenting the App notifications component. Additionally, the commit adds inline docs in the useAppNotifications hook.

* feat: move edit event route